+++ plugins/board-resources/src/components/editor/CardDetails.svelte
@@ -15,21 +15,54 @@
 <script lang="ts">
   import type { Card, CardDate } from '@anticrm/board'
 
+  import contact, { Employee } from '@anticrm/contact'
   import { getResource } from '@anticrm/platform'
-  import { getClient } from '@anticrm/presentation'
-  import { Label } from '@anticrm/ui'
+  import { createQuery, getClient } from '@anticrm/presentation'
+  import { Button, IconAdd, Label } from '@anticrm/ui'
 
   import board from '../../plugin'
   import { getCardActions } from '../../utils/CardActionUtils'
-  import { hasDate } from '../../utils/CardUtils'
+  import { hasDate, updateCardMembers } from '../../utils/CardUtils'
   import DatePresenter from '../presenters/DatePresenter.svelte'
-  import MembersPresenter from '../presenters/MembersPresenter.svelte'
+  import MemberPresenter from '../presenters/MemberPresenter.svelte'
   import CardLabels from './CardLabels.svelte'
 
   export let value: Card
+  const query = createQuery()
   const client = getClient()
+  let members: Employee[]
   let membersHandler: (e: Event) => void
   let dateHandler: (e: Event) => void
+
+  $: membersIds = members?.map((m) => m._id) ?? []
+
+  const getMenuItems = (member: Employee) => {
+    return [
+      [
+        {
+          title: board.string.ViewProfile,
+          handler: () => console.log('TODO: implement')
+        }
+      ],
+      [
+        {
+          title: board.string.RemoveFromCard,
+          handler: () => {
+            const newMembers = membersIds.filter((m) => m !== member._id)
+            updateCardMembers(value, client, newMembers)
+          }
+        }
+      ]
+    ]
+  }
+
+  $: if (value.members && value.members.length > 0) {
+    query.query(contact.class.Employee, { _id: { $in: value.members } }, (result) => {
+      members = result
+    })
+  } else {
+    members = []
+  }
 
   function updateDate (e: CustomEvent<CardDate>) {
     client.update(value, { date: e.detail })
@@ -52,12 +85,17 @@
 </script>
 
 {#if value}
-  {#if value.members?.length}
+  {#if members && members.length > 0}
     <div class="flex-col mt-4 mr-6">
       <div class="text-md font-medium">
         <Label label={board.string.Members} />
       </div>
-      <MembersPresenter object={value} {membersHandler} />
+      <div class="flex-row-center flex-gap-1">
+        {#each members as member}
+          <MemberPresenter value={member} size="large" menuItems={getMenuItems(member)} />
+        {/each}
+        <Button icon={IconAdd} shape="circle" kind="no-border" size="large" on:click={membersHandler} />
+      </div>
     </div>
   {/if}
   {#if value.labels && value.labels.length > 0}
